% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{build_adjacency_mat}
\alias{build_adjacency_mat}
\title{Threshold to adjacency matrix}
\usage{
build_adjacency_mat(cor_mat, cor_threshold = 0.7)
}
\arguments{
\item{cor_mat}{The gene-gene correlation matrix}

\item{cor_threshold}{Correlation threshold for defining edges (default: 0.7)}
}
\value{
A binary adjacency matrix with 1 indicating an edge
and 0 indicating no edge
}
\description{
Converts gene-gene correlation matrix into a binary adjacency matrix
according to threshold for absolute correlation value.
}
\examples{
data(example_se)
mat <- prepare_expression_matrix(example_se)
cor_mat <- pairwise_corr(mat)
adj_mat <- build_adjacency_mat(cor_mat, cor_threshold=0.7)
}
