% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{get_hub_genes}
\alias{get_hub_genes}
\title{Find and subset hub genes}
\usage{
get_hub_genes(network, modules = NULL, n = 20)
}
\arguments{
\item{network}{A network object output by build_network()}

\item{modules}{A named integer vector of module assignments}

\item{n}{Number of top hub genes to return (default: 20)}
}
\value{
A data frame of hub genes ranked by degree (highest to lowest)
}
\description{
Rank genes by degree and extract top \code{n} hub genes from network
}
\examples{
data(example_se)
mat <- prepare_expression_matrix(example_se)
cor_mat <- pairwise_corr(mat)
adj_mat <- build_adjacency_mat(cor_mat)
net <- build_network(adj_mat)
gene_mods <- detect_network_modules(net)
hub_genes <- get_hub_genes(net, modules = gene_mods$modules, n = 20)
}
