% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toast.R
\name{make_toast}
\alias{make_toast}
\title{Make toast with any bread you want}
\usage{
make_toast(bread, buttered = FALSE)
}
\arguments{
\item{bread}{The type of bread you want to use for your toast}

\item{buttered}{A boolean value indicating whether you want your toast buttered or not}
}
\value{
A string describing the toast you made
}
\description{
Make toast with any bread you want
}
\examples{
make_toast("sourdough", buttered = TRUE)
}
\author{
Jared Andrews
}
