% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_corr_map}
\alias{plot_corr_map}
\title{Plot gene-gene correlation heatmap}
\usage{
plot_corr_map(cor_mat, cor_method = "pearson")
}
\arguments{
\item{cor_mat}{A gene-gene correlation matrix}

\item{cor_method}{Correlation method used to compute matrix
(default: "pearson")}
}
\value{
Renders the heatmap to the active graphics device.
The ComplexHeatmap object is returned invisibly for optional reuse.
}
\description{
Plot gene-gene correlation heatmap
}
\examples{
data(example_se)
mat <- prepare_expression_matrix(example_se)
cor_mat <- pairwise_corr(mat)
ht <- plot_corr_map(cor_mat, cor_method = "pearson")
}
