% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_variance_explained}
\alias{plot_variance_explained}
\title{Plot variance explained by principal components}
\usage{
plot_variance_explained(pca_result, n_pcs = 8)
}
\arguments{
\item{pca_result}{Output from \code{\link{run_pca}()}.}

\item{n_pcs}{Maximum number of PCs to display. Default: 8.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object.
}
\description{
Produces a bar chart showing the percentage of variance explained by each
principal component, using the output of \code{\link{run_pca}()}.
}
\examples{
data(airway, package = "airway")
result <- run_pca(airway, n_top = 50)
plot_variance_explained(result)
}
\author{
Jared Andrews
}
