% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-prep.R
\name{select_random_samples}
\alias{select_random_samples}
\title{Select set of random samples}
\usage{
select_random_samples(se, n = 200)
}
\arguments{
\item{se}{A SummarizedExperiment object}

\item{n}{Number of random samples to select (default: 200)}
}
\value{
A SummarizedExperiment subset to n random samples
}
\description{
Select set of random samples
}
\note{
For reproducibility, set a random seed with \code{set.seed()}
before calling this function.
}
