% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{summarize_network}
\alias{summarize_network}
\title{Summarize global network properties}
\usage{
summarize_network(network, modules, modularity_score)
}
\arguments{
\item{network}{A network object output by build_network()}

\item{modules}{A named integer vector of module assignments}

\item{modularity_score}{A numeric modularity score}
}
\value{
A data frame of global network statistics
}
\description{
Summarize global network properties
}
\examples{
data(example_se)
mat <- prepare_expression_matrix(example_se)
cor_mat <- pairwise_corr(mat)
adj_mat <- build_adjacency_mat(cor_mat)
net <- build_network(adj_mat)
gene_mods <- detect_network_modules(net)
global_net_summary <- summarize_network(net, gene_mods$modules,
    gene_mods$modularity)
}
