% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-prep.R
\name{top_variable_features}
\alias{top_variable_features}
\title{Select top variable features}
\usage{
top_variable_features(se, n = 500, assay_name = "counts")
}
\arguments{
\item{se}{A SummarizedExperiment object}

\item{n}{Number of top variable features to select (default: 500)}

\item{assay_name}{Name of assay to use (default: "counts")}
}
\value{
A SummarizedExperiment subset to the top n variable features
}
\description{
Select top variable features
}
\examples{
# Assuming 'se' is a SummarizedExperiment
data(example_se)
se_top <- top_variable_features(example_se, n = 500)
}
